+++ spec/sensitive_data_filter/middleware/occurrence_spec.rb
@@ -7,7 +7,8 @@ require 'sensitive_data_filter/middleware/occurrence'
 describe SensitiveDataFilter::Middleware::Occurrence do
   let(:ip) { '127.0.0.1' }
   let(:request_method) { 'POST' }
-  let(:url) { 'https://test.example.com.au/test' }
+  let(:original_url) { 'https://test.example.com.au/test?credit_card=4111111111111111' }
+  let(:filtered_url) { 'https://test.example.com.au/test?credit_card=%5BFILTERED%5D' }
   let(:content_type) { 'application/json' }
   let(:original_query_params) { { 'credit_card' => '4111 1111 1111 1111' } }
   let(:filtered_query_params) { { 'credit_card' => '[FILTERED]' } }
@@ -18,7 +19,7 @@ describe SensitiveDataFilter::Middleware::Occurrence do
     double(
       ip:             ip,
       request_method: request_method,
-      url:            url,
+      url:            original_url,
       content_type:   content_type,
       query_params:   original_query_params,
       body_params:    original_body_params,
@@ -29,7 +30,7 @@ describe SensitiveDataFilter::Middleware::Occurrence do
     double(
       ip:             ip,
       request_method: request_method,
-      url:            url,
+      url:            filtered_url,
       content_type:   content_type,
       query_params:   filtered_query_params,
       body_params:    filtered_body_params,
@@ -53,7 +54,7 @@ describe SensitiveDataFilter::Middleware::Occurrence do
   specify { expect(occurrence.matches).to eq matches }
   specify { expect(occurrence.origin_ip).to eq ip }
   specify { expect(occurrence.request_method).to eq request_method }
-  specify { expect(occurrence.url).to eq url }
+  specify { expect(occurrence.url).to eq filtered_url }
   specify { expect(occurrence.original_query_params).to eq original_query_params }
   specify { expect(occurrence.original_body_params).to eq original_body_params }
   specify { expect(occurrence.filtered_query_params).to eq filtered_query_params }
@@ -65,7 +66,7 @@ describe SensitiveDataFilter::Middleware::Occurrence do
     {
       origin_ip:             ip,
       request_method:        request_method,
-      url:                   url,
+      url:                   filtered_url,
       content_type:          content_type,
       filtered_query_params: filtered_query_params,
       filtered_body_params:  filtered_body_params,
@@ -78,7 +79,7 @@ describe SensitiveDataFilter::Middleware::Occurrence do
     "[SensitiveDataFilter] Sensitive Data detected and masked:\n"\
     "Origin Ip: 127.0.0.1\n"\
     "Request Method: POST\n"\
-    "Url: https://test.example.com.au/test\n"\
+    "Url: https://test.example.com.au/test?credit_card=%5BFILTERED%5D\n"\
     "Content Type: application/json\n"\
     "Filtered Query Params: {\"credit_card\"=>\"[FILTERED]\"}\n"\
     "Filtered Body Params: {:credit_cards=>\"[FILTERED] and [FILTERED]\"}\n"\