+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 
+## [Unreleased]
+### Fixed
+- Ensures that the url returned by the occurrence is filtered
+
 ## [0.2.2] - 2016-12-21
 ### Fixed
 - Implements stricter credit cards pattern matching